+++ src/features/editHelpers.ts
@@ -11,22 +11,34 @@ export function loadEditHelpers(subscriptions: Disposable[]) {
   subscriptions.push(
     languages.setLanguageConfiguration("logtalk", {
       indentationRules: {
-        decreaseIndentPattern: /(\s*\)|\s*\])$/,
-        increaseIndentPattern: /(.*:-\s*|.*->\s*|.+\[|.+\()$/
+        decreaseIndentPattern: /(?:)/,
+        increaseIndentPattern: /(?:)/
       },
       wordPattern: /(-?\d*\.\d\w*)|([^\`\~\!\@\#\%\^\&\*\(\)\-\=\+\[\{\]\}\\\|\;\:\'\"\,\.\<\>\/\?\s]+)/g,
       onEnterRules: [
         {
-          beforeText: /(^\s*|.*%.+|^\s+.+\.)$/,
+          beforeText: /(^\s*|.*%.+)$/,
           action: { indentAction: IndentAction.None }
         },
         {
-          beforeText: /.+:-|:- object.+\.|:- protocol.+\.|:- category.+\.|.+\([^\)]*$/,
+          beforeText: /.*:-$/,
           action: { indentAction: IndentAction.Indent }
         },
         {
-          beforeText: /(.+\.|[^,;])$/,
+          beforeText: /.*-->$/,
+          action: { indentAction: IndentAction.Indent }
+        },
+        {
+          beforeText: /.+\.$/,
           action: { indentAction: IndentAction.Outdent }
+        },
+        {
+          beforeText: /.*->$/,
+          action: { indentAction: IndentAction.Indent }
+        },
+        {
+          beforeText: /:- object\(.+\.|:- protocol\(.+\.|:- category\(.+\.|.+\([^\)]*$/,
+          action: { indentAction: IndentAction.Indent }
         },
         {
           // e.g. /** | */