+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [0.22.0]
 
 * Change linters to no longer create and write to an "OUTPUT" pane channel
+* Show current clause in the active text editor when debugging
 * Add Logtalk source file icons
 * Add "Logtalk: Make - Circular" command
 * Add "Logtalk: Make - Optimal" command